+++ MovilesProject/src/main/java/com/moviles/model/entity/Post.java
@@ -20,7 +20,7 @@ public abstract class Post {
     @Field("emailUsuario")
     private String userName;
     @Field("referencia")
-    private int referencia;
+    private Long referencia;
     @Field("idMovilVenta")
     private Long referenciaMovil;
     @Field("estado")
@@ -28,7 +28,7 @@ public abstract class Post {
     @Field("fechaCreacion")
     private Date date;
 
-    public Post(String username, int referencia,Long referenciaMovil, String estado, Date date) {
+    public Post(String username, Long referencia,Long referenciaMovil, String estado, Date date) {
         this.userName = username;
         this.referencia = referencia;
         this.referenciaMovil = referenciaMovil;
