+++ MovilesProject/src/main/java/com/moviles/model/entity/PostIntercambio.java
@@ -17,7 +17,7 @@ public class PostIntercambio extends Post {
     @Field("estadoMinimo")
     private String estadoMinimo;
 
-    public PostIntercambio(String emailUsuario, int referencia,Long referenciaMovil,  String estado, Date date,
+    public PostIntercambio(String emailUsuario, Long referencia,Long referenciaMovil,  String estado, Date date,
             String estadoMinimo) {
         super(emailUsuario, referencia,referenciaMovil, estado, date);
         this.estadoMinimo = estadoMinimo;
